cli : reorg p1
